pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "git-autograder"
version = "4.0.0"
authors = [{ name = "Jiahao, Woo", email = "woojiahao1234@gmail.com" }]
description = "Library for autograding Git repositories"
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Topic :: Software Development",
  "Programming Language :: Python :: 3.13",
]
license.file = "LICENSE"
dependencies = [
  "pytz",
  "types-pytz",
  "difflib_parser",
  "GitPython",
  "repo-smith",
]

[project.urls]
Homepage = "https://github.com/git-mastery/git-autograder.git"
Repository = "https://github.com/git-mastery/git-autograder.git"
Issues = "https://github.com/git-mastery/git-autograder/issues"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
pythonpath = ["src"]

[tool.mypy]
mypy_path = "src"
